Final submitted version, not working properly
here is an email I sent about the project:

Hi [TA], Unfortunately I was unable to get the A* path finding algorithm to work correctly within the time given, so the program cannot find any valid paths. The part of the program that is working fully is reading from files and parsing the data in preparation for the algorithm. (you can either rename the second test surface file to `testSurface1.txt` or change it in the construction: ` World mountain("testSurface1.txt"); ` in main to see it work with 2 different files, and if you want more file tests change the equation in `surface generator.py` and run that to generate another surface).

I have a point class with data members height and two scores needed for the path finding algorithm, and the constructors, a validity check, and the relevant getters and setters.

The readSurface() function in World.cpp reads all the height values in the file, changing height 0 to 1 (to allow a zero height to be reserved as invalid to save memory, as the full surface files will be about 4GB large) then placing those heights into a 2D vector of my point class, also setting the cost scores to a default of infinity (the largest positive integer, required for the path finding algorithm if it was working). This 2D vector is stored in the `_surface` data member of the World class. Then in main it displays the heights in this 2D vector to the console (I was going to display the surface and draw the best path onto it in an SFML window but ran out of time).

Also, my apologies for including too many files in the submission. You don't need SFML, and you can remove the entire `function` folder and the `Map.h` file in the class folder, these were files I did some testing in before adding them to the World class. I included the `CMakeLists.txt` to show which files should be included in the build.